+++ packages/spacecat-shared-ahrefs-client/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.7](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.6...@adobe/spacecat-shared-ahrefs-client-v1.6.7) (2025-02-16)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#603](https://github.com/adobe/spacecat-shared/issues/603)) ([b58d4c7](https://github.com/adobe/spacecat-shared/commit/b58d4c7237fb2522bba9b722e9eed7b0ae9e5f70))
+
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.5...@adobe/spacecat-shared-ahrefs-client-v1.6.6) (2025-02-12)
+
+
+### Bug Fixes
+
+* **ahrefs-client:** enhance organic keywords query to use in HOTLCTR opportunity ([#589](https://github.com/adobe/spacecat-shared/issues/589)) ([23aef2e](https://github.com/adobe/spacecat-shared/commit/23aef2eef7a2e44b3d68ec8ba64378e8d4a3b605))
+
+# [@adobe/spacecat-shared-ahrefs-client-v1.6.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.4...@adobe/spacecat-shared-ahrefs-client-v1.6.5) (2025-02-08)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#587](https://github.com/adobe/spacecat-shared/issues/587)) ([14cce0a](https://github.com/adobe/spacecat-shared/commit/14cce0aa900b4a1b3bbec2d48e6d37766c7769ee))
+
 # [@adobe/spacecat-shared-ahrefs-client-v1.6.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.6.3...@adobe/spacecat-shared-ahrefs-client-v1.6.4) (2025-01-16)